Renamed, updated mapper, readme
