+++ .travis.yml
@@ -15,8 +15,9 @@ language:
 python:
   - 2.7
   - 3.3
-  - 3.4
+  # - 3.4
   - 3.5
+  - 3.6
 
 env:
   global:
