+++ tableschema_bigquery/mapper.py
@@ -1,0 +1,141 @@
+# -*- coding: utf-8 -*-
+from __future__ import division
+from __future__ import print_function
+from __future__ import absolute_import
+from __future__ import unicode_literals
+
+import re
+from slugify import slugify
+
+
+# Module API
+
+class Mapper(object):
+
+    # Public
+
+    def __init__(self, prefix):
+        """Mapper to convert/restore FD entities to/from BigQuery entities
+        """
+        self.__prefix = prefix
+
+    def convert_bucket(self, bucket):
+        """Convert bucket to BigQuery
+        """
+        return self.__prefix + bucket
+
+    def convert_descriptor(self, descriptor):
+        """Convert descriptor to BigQuery
+        """
+
+        # Convert
+        fields = []
+        for field in descriptor['fields']:
+            ftype = self.convert_type(field['type'])
+            mode = 'NULLABLE'
+            if field.get('constraints', {}).get('required', False):
+                mode = 'REQUIRED'
+            fields.append({
+                'name': _slugify_field_name(field['name']),
+                'type': ftype,
+                'mode': mode,
+            })
+        nativedesc = {'fields': fields}
+
+        return nativedesc
+
+    def convert_row(self, row):
+        """Convert row to BigQuery
+        """
+        return row
+
+    def convert_type(self, type):
+        """Convert type to BigQuery
+        """
+
+        # Mapping
+        mapping = {
+            'string': 'STRING',
+            'integer': 'INTEGER',
+            'number': 'FLOAT',
+            'boolean': 'BOOLEAN',
+            'date': 'TIMESTAMP',
+            'time': 'TIMESTAMP',
+            'datetime': 'TIMESTAMP',
+        }
+
+        # Not supported type
+        if type not in mapping:
+            message = 'Type %s is not supported' % type
+            raise TypeError(message)
+
+        return mapping[type]
+
+    def restore_bucket(self, table_name):
+        """Restore bucket from BigQuery
+        """
+        if table_name.startswith(self.__prefix):
+            return table_name.replace(self.__prefix, '', 1)
+        return None
+
+    def restore_descriptor(self, nativedesc):
+        """Restore descriptor rom BigQuery
+        """
+
+        # Convert
+        fields = []
+        for field in nativedesc['fields']:
+            ftype = self.restore_type(field['type'])
+            resfield = {
+                'name': field['name'],
+                'type': ftype,
+            }
+            if field.get('mode', 'NULLABLE') != 'NULLABLE':
+                resfield['constraints'] = {'required': True}
+            fields.append(resfield)
+        descriptor = {'fields': fields}
+
+        return descriptor
+
+    def restore_row(self, row):
+        """Restore row from BigQuery
+        """
+        return row
+
+    def restore_type(self, type):
+        """Restore type from BigQuery
+        """
+
+        # Mapping
+        mapping = {
+            'STRING': 'string',
+            'INTEGER': 'integer',
+            'FLOAT': 'number',
+            'BOOLEAN': 'boolean',
+            'TIMESTAMP': 'datetime',
+        }
+
+        # Not supported type
+        if type not in mapping:
+            message = 'Type %s is not supported' % type
+            raise TypeError(message)
+
+        return mapping[type]
+
+
+# Internal
+
+def _slugify_field_name(name):
+
+    # Referene:
+    # https://cloud.google.com/bigquery/docs/reference/v2/tables
+    MAX_LENGTH = 128
+    VALID_NAME = '^[a-zA-Z_]\w{0,%d}$' % (MAX_LENGTH-1)
+
+    # Convert
+    if not re.match(VALID_NAME, name):
+        name = slugify(name, separator='_')
+        if not re.match('^[a-zA-Z_]', name):
+            name = '_' + name
+
+    return name[:MAX_LENGTH]
