+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-package=jsontableschema_bigquery
+package=tableschema_bigquery
 skip_missing_interpreters=true
 envlist=
   py27
